+++ all_files.vcxitems
@@ -98,7 +98,6 @@
     <ClCompile Include="$(MSBuildThisFileDirectory)cdc_rsync\client_socket.cc" />
     <ClCompile Include="$(MSBuildThisFileDirectory)cdc_rsync\file_finder_and_sender.cc" />
     <ClCompile Include="$(MSBuildThisFileDirectory)cdc_rsync\file_finder_and_sender_test.cc" />
-    <ClCompile Include="$(MSBuildThisFileDirectory)cdc_rsync\cdc_rsync.cc" />
     <ClCompile Include="$(MSBuildThisFileDirectory)cdc_rsync\cdc_rsync_client.cc" />
     <ClCompile Include="$(MSBuildThisFileDirectory)cdc_rsync\parallel_file_opener.cc" />
     <ClCompile Include="$(MSBuildThisFileDirectory)cdc_rsync\parallel_file_opener_test.cc" />
@@ -201,9 +200,7 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)cdc_rsync\base\socket.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)cdc_rsync\client_file_info.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)cdc_rsync\client_socket.h" />
-    <ClInclude Include="$(MSBuildThisFileDirectory)cdc_rsync\error_messages.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)cdc_rsync\file_finder_and_sender.h" />
-    <ClInclude Include="$(MSBuildThisFileDirectory)cdc_rsync\cdc_rsync.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)cdc_rsync\cdc_rsync_client.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)cdc_rsync\parallel_file_opener.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)cdc_rsync\progress_tracker.h" />
